Improve graph performance
Based on my testing, this restores performance to where it was for v0.2 with the exception of the last-month graphs while still generating the same data as v0.7.  Excluding last month graphs, v0.7 tended to be 15-35% worse than v0.7 in performance.  The new code is comparable to v0.2, again excluding last-month graphs, with the performance either being equivalent or up to 10% worse depending on the particular deck and what is being graphed.  The performance of the last-month graphs in the new version is still worse but I don't see any way to avoid this given how Anki is tracking the reviews.  The type (learning, relearning, cram, etc.) can't be trusted, so you have to query past history in order to get accurate data.  The effect of doing this is obviously worst for last-month graphs.  Still, the performance for last-month graphs in the new version tends to be better than that of v0.7.