Fix mature card calculations
The cumulative sum of net matured cards didn't match the total number of matured cards when viewed over deck lifetime.  Apparently there is or was some issue with lastIvl not being updated correctly.  To work around this it isn't used and instead the previous value of ivl is recorded.  This fixes the issue.